requirements.txt

streamlit
requests
google-api-python-client
google-auth
google-auth-oauthlib
